+++ src/test/java/de/codescape/jira/plugins/scrumpoker/action/ErrorLogActionTest.java
@@ -20,7 +20,7 @@ import static org.hamcrest.Matchers.is;
 import static org.junit.Assert.assertThat;
 import static org.mockito.Mockito.*;
 
-public class ShowErrorLogActionTest {
+public class ErrorLogActionTest {
 
     @Rule
     public MockitoContainer mockitoContainer = MockitoMocksInContainer.rule(this);
@@ -33,7 +33,7 @@ public class ShowErrorLogActionTest {
     private ErrorLogService errorLogService;
 
     @InjectMocks
-    private ShowErrorLogAction showErrorLogAction;
+    private ErrorLogAction errorLogAction;
 
     @Mock
     private HttpServletRequest httpServletRequest;
@@ -42,22 +42,22 @@ public class ShowErrorLogActionTest {
     public void shouldExposeTheErrorList() {
         List<ScrumPokerError> scrumPokerErrors = new ArrayList<>();
         when(errorLogService.listAll()).thenReturn(scrumPokerErrors);
-        assertThat(showErrorLogAction.getErrorList(), is(equalTo(scrumPokerErrors)));
+        assertThat(errorLogAction.getErrorList(), is(equalTo(scrumPokerErrors)));
     }
 
     @Test
     public void shouldEmptyTheErrorLogWhenRequested() {
         when(httpServletVariables.getHttpRequest()).thenReturn(httpServletRequest);
-        when(httpServletRequest.getParameter(ShowErrorLogAction.Parameters.ACTION)).thenReturn("empty");
-        assertThat(showErrorLogAction.doExecute(), is(equalTo(showErrorLogAction.SUCCESS)));
+        when(httpServletRequest.getParameter(ErrorLogAction.Parameters.ACTION)).thenReturn("empty");
+        assertThat(errorLogAction.doExecute(), is(equalTo(errorLogAction.SUCCESS)));
         verify(errorLogService, times(1)).emptyErrorLog();
     }
 
     @Test
     public void shouldNotEmptyTheErrorWhenNotRequested() {
         when(httpServletVariables.getHttpRequest()).thenReturn(httpServletRequest);
-        when(httpServletRequest.getParameter(ShowErrorLogAction.Parameters.ACTION)).thenReturn(null);
-        assertThat(showErrorLogAction.doExecute(), is(equalTo(showErrorLogAction.SUCCESS)));
+        when(httpServletRequest.getParameter(ErrorLogAction.Parameters.ACTION)).thenReturn(null);
+        assertThat(errorLogAction.doExecute(), is(equalTo(errorLogAction.SUCCESS)));
         verify(errorLogService, never()).emptyErrorLog();
     }